Rename scatterplot functions
