+++ src/utils/plotSpecs/generateMitochondrialFractionScatterplot.js
